Move package and inventory handlers to own file
